Changed default keys
Made default keys easier to use. The keys are concentrated close to
WASD.

Walking changed from arrow-keys to W, A and D.
Rope-swing changed to S.
Aiming is still done with up and down arrows.
Changin weapons is done with Q and E.

Shooting is temporarily done with F, however the resault was not very
good.